refactor map code
